// src/types/configs/global-configs.types.ts
type themeType = {
  prefix: string | { idle: string; done: string };
  spinner: {
    interval: number;
    frames: string[];
  };
  style: {
    answer: (text: string) => string;
    message: (text: string, status: string) => string;
    error: (text: string) => string;
    defaultAnswer: (text: string) => string;
    help: (text: string) => string;
    highlight: (text: string) => string;
    key: (text: string) => string;
    disabled: (text: string) => string;
    description: (text: string) => string;
    keysHelpTip: (keys: [key: string, action: string][]) => string | undefined;
  };
  icon: {
    cursor: string;
  };
  helpMode: "always" | "never" | "auto";
  indexMode: "hidden" | "number";
  validationFailureMode: "keep" | "clear";
};

export { themeType };